requirements.txt
# PyTorch is already in the container base image (pytorch/pytorch:2.4.0-cuda12.4).
# If installing locally:
#   pip install torch --index-url https://download.pytorch.org/whl/cu124   (GPU)
#   pip install torch --index-url https://download.pytorch.org/whl/cpu     (CPU)
torch>=2.0

sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.65

# flash-attn speeds up 8B models significantly (Qwen3-8B, Nemotron-8B).
# Requires CUDA and a build toolchain — installed separately in container.def.
# Uncomment for local GPU install:
# flash-attn>=2.6.0
